replace add liquidty trait
