+++ traits/src/lib.rs
@@ -32,7 +32,7 @@ pub use oracle::*;
 pub use registry::*;
 
 use codec::{Decode, Encode};
-use frame_support::dispatch::{self};
+use frame_support::dispatch::{self, DispatchResult};
 use frame_support::sp_runtime::{traits::Zero, DispatchError, RuntimeDebug};
 use frame_support::traits::LockIdentifier;
 use frame_support::weights::Weight;
@@ -274,6 +274,17 @@ pub trait AMMPosition<AssetId, Balance> {
 		asset_b: AssetId,
 		shares_amount: Balance,
 	) -> Result<(Balance, Balance), Self::Error>;
+}
+
+/// Implementers of this trait are able to add liquidity to their AMM pool.
+pub trait XykAddLiquidity<Origin, AssetId, Balance> {
+	fn add_liquidity(
+		origin: Origin,
+		asset_a: AssetId,
+		asset_b: AssetId,
+		amount_a: Balance,
+		amount_b_max_limit: Balance,
+	) -> DispatchResult;
 }
 
 /// Provides account's fee payment asset
